requirements.txt
ollama
chromadb
llama-index
litellm
groq
docx2txt
colorama
termcolor
click
asciitree
fastapi
weave
agentops
langchain
langchain_core
watchdog
agentops
python-dotenv